First refinement update
minor changes:
Streamlined the inputs for the learning model.
Added extensive comments for clarity
Delegated the device setting to the model instead of the dataset object
Added a dependency on torchinfo (still unused)
Avoided keeping the whole dataset as a variable in the learning model
Changed syntax in the learning model to use only the local variables, rather than dataset variables.